[mason] Simplify durable runtime SDK
