+++ integrations/mason/pyproject.toml
@@ -40,6 +40,7 @@ runtime = [
     "uvicorn>=0.20.0",
 ]
 runtime-openai = [
+    "databricks-ai-bridge[memory]>=0.21.0",
     "openai-agents>=0.7.0",
     "databricks-openai>=0.13.0",
     "fastapi>=0.129.0",
@@ -47,6 +48,7 @@ runtime-openai = [
     "uuid-utils>=0.10.0",
     "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
     "databricks-agents>=1.9.3",
+    "uvicorn>=0.20.0",
 ]
 
 [project.scripts]
