+++ integrations/mason/src/databricks_mason/durable_server/__init__.py
@@ -1,0 +1,5 @@
+"""HTTP application for Mason durable agent execution."""
+
+from databricks_mason.durable_server.app import DurableAgentApp, DurableAgentContext
+
+__all__ = ["DurableAgentApp", "DurableAgentContext"]
