+++ integrations/mason/templates/agent-langgraph/.env.example
@@ -34,12 +34,3 @@ DATABRICKS_CONFIG_PROFILE=DEFAULT
 # durable graph state incl. paused HITL runs, over RPCs (no DB connection). Access uses your normal
 # Databricks auth (must have access to the store). See databricks_mason/langgraph/session_store.py.
 # AGENT_SESSION_STORE=my-agent-sessions
-
-# --- Optional: Lakebase-backed runtime durability for local development ---
-# `mason deploy` configures this automatically: it reuses the Session Store database, otherwise the
-# Memory Store database, otherwise a dedicated <app>-durability project. For an explicit local test,
-# set either an endpoint resource path or a project + branch. The runtime adds only its
-# `databricks_mason_runtime` schema and tables to that database.
-# LAKEBASE_AUTOSCALING_ENDPOINT=projects/<project>/branches/<branch>/endpoints/<endpoint>
-# LAKEBASE_AUTOSCALING_PROJECT=<project>
-# LAKEBASE_AUTOSCALING_BRANCH=<branch>
