+++ integrations/mason/templates/agent-langgraph/agent/mcps.py
@@ -1,7 +1,8 @@
 """MCP servers to offer the agent — this is where you configure them.
 
-Empty by default: the agent runs with no MCP servers. Add servers to ``build_mcp_servers`` to offer
-them; ``databricks_mason.langgraph.mcp`` turns them into tools for each request.
+Empty by default: the agent runs with only the MCP servers declared in ``agent.toml``. Add servers
+to ``build_mcp_servers`` to offer more; ``agent.py`` passes them to ``mcp_tools``, which joins them
+with the ``agent.toml`` servers and fetches their tools each request.
 """
 
 from databricks_langchain import DatabricksMCPServer
