+++ integrations/mason/templates/agent-langgraph/app.yaml
@@ -1,12 +1,10 @@
 command: ["uv", "run", "start-server"]
 # databricks apps listen by default on port 8000
 #
-# No env vars are required by default: the app runs the lean backend (in-process checkpointer,
-# tracing off). `mason deploy` separately attaches one Lakebase database for runtime invocation
-# durability and injects its internal endpoint setting. It reuses Session Store first, Memory Store
-# second, or provisions a dedicated <app>-durability project. Do not add that internal setting here.
-# To enable tracing, long-term memory, or durable LangGraph checkpoints, add the matching entries
-# below (literal values):
+# No env vars are required by default: local development uses in-process state. `mason deploy`
+# attaches one Lakebase database for runtime durability, reusing Session Store first, Memory Store
+# second, or provisioning a dedicated <app>-durability project. To enable tracing, long-term memory,
+# or durable LangGraph checkpoints, add the matching entries below (literal values):
 #
 # env:
 #   # Tracing (set a destination + an experiment). Destination: MLFLOW_TRACKING_URI or
