+++ integrations/mason/templates/agent-openai/.env.example
@@ -2,7 +2,7 @@
 #
 # The only required setting for local dev is a Databricks auth profile (used to call the model).
 # Everything else — MLflow tracing, long-term memory — is optional and off by default. Conversation
-# history uses an in-process session (SQLite in :memory:) with no config; durability is one env var.
+# history and runtime invocation state use in-process stores with no config.
 
 # Databricks auth profile (from `databricks auth profiles`). Used to call the model endpoint.
 DATABRICKS_CONFIG_PROFILE=DEFAULT
