+++ integrations/mason/templates/agent-openai/app.yaml
@@ -1,9 +1,10 @@
 command: ["uv", "run", "start-server"]
 # databricks apps listen by default on port 8000
 #
-# No env vars are required by default: the app runs the lean backend (in-process session, tracing
-# off). To enable tracing, long-term memory, or durable state, add the matching entries below
-# (literal values):
+# No env vars are required by default: local development uses in-process state. `mason deploy`
+# attaches one Lakebase database for runtime durability, reusing Session Store first, Memory Store
+# second, or provisioning a dedicated <app>-durability project. To enable tracing, long-term memory,
+# or managed session history, add the matching entries below (literal values):
 #
 # env:
 #   # Tracing (set a destination + an experiment). Destination: MLFLOW_TRACKING_URI or
