+++ integrations/mason/templates/agent-openai/runtime/main.py
@@ -1,27 +1,30 @@
-"""Agent server entry point.
-
-Loads config, wires the agent's handlers into the (SDK-agnostic) FastAPI app from
-``runtime/runtime.py``, and runs uvicorn. The handlers live in ``agent/agent.py`` — the only
-SDK-specific piece.
-"""
+"""Run the SDK-hosted durable agent application."""
 
 import os
 from pathlib import Path
 
-# Importing the agent is side-effect-free (no env is read until configure()), so it sits up top.
 import agent.agent
 import uvicorn
+from databricks_mason import DurableAgentApp
+from databricks_mason.runtime.store import InMemoryDurabilityStore, LakebaseDurabilityStore
 from dotenv import load_dotenv
 
-from runtime.runtime import build_app
-
-# Load .env before configure() reads env (agent client auth + tracing config), then wire the agent.
 load_dotenv(dotenv_path=Path(__file__).parent.parent / ".env", override=True)
 agent.agent.configure()
 
-# Module-level app so uvicorn can import it by string (and to enable multiple workers).
-app = build_app(agent.agent.invoke_handler, agent.agent.stream_handler)
+endpoint = os.getenv("DATABRICKS_MASON_RUNTIME_ENDPOINT")
+durability_store = (
+    LakebaseDurabilityStore.from_app_resource(endpoint=endpoint)
+    if endpoint
+    else InMemoryDurabilityStore()
+)
+server = DurableAgentApp(
+    agent.agent.invoke,
+    on_resume=agent.agent.recover,
+    durability_store=durability_store,
+)
+app = server.app
 
 
-def main():
+def main() -> None:
     uvicorn.run("runtime.main:app", host="0.0.0.0", port=int(os.getenv("PORT", "8000")))
